🔧 Fix type casting and add bonding curve test
- Fix PrimarySaleV3: Cast int256 coordinates to uint256 for mintParcel
- Add TestBondingCurve script to verify pricing math
- Test all three phases of bonding curve (0-50, 50-200, 200+)

+++ contracts/src/PrimarySaleV3.sol
@@ -239,10 +239,11 @@ contract PrimarySaleV3 is Ownable, ReentrancyGuard {
         totalETHCollected += msg.value;
 
         // Mint parcel via HyperLandCore (primary sale)
+        // Cast int256 coordinates to uint256 for mintParcel
         hyperLandCore.mintParcel(
             msg.sender,
-            config.x,
-            config.y,
+            uint256(config.x),
+            uint256(config.y),
             config.size,
             config.assessedValue
         );
